+++ packages/site/src/routes/+page.svelte
@@ -1,2 +1,5 @@
-<h1>Welcome to SvelteKit</h1>
-<p>Visit <a href="https://kit.svelte.dev">kit.svelte.dev</a> to read the documentation</p>
+<p>
+  The only family information available at this time is <a href="gary"
+    >Gary's</a
+  >
+</p>