+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+# 14 Mar 2023
+
+- Switch to svelte-kit for real
+
 # 5 May 2022
 
 - Switch to svelte-kit